Add grunt to the project
This adds a new command to the project `grunt lint`. This will run
jshint on the files in the js/actions and js/main.js file using the
rules found in the .jshintrc file.

I have a very strict .jshintrc file that I'd like to share in the
future, but this is a good starting point for now.

This adds an npm package.json file to the project. I prefer it to be be
canonical source of information about the project, and I tend to
relegate bower.json to only have a list of browser dependencies.

You'll have to install the grunt-cli on your system if you haven't done
so before and then run and `npm install` before you run `grunt`.